Final Bug Fixes after Testing
You might encounter some errors if trying to test these with an old database (because of the indexing).
Don't be alarmed, my suggestion is to remove all data, when finding these errors, from the database (either manually or from the software) and then try testing again. The errors should disappear.